refactor(editor): Extract collaboration code into features (no-changelog)
