+++ packages/frontend/editor-ui/src/features/collaboration/components/CollaborationPane.test.ts
@@ -2,7 +2,7 @@ import { createTestingPinia } from '@pinia/testing';
 import { mock } from 'vitest-mock-extended';
 
 import { STORES } from '@n8n/stores';
-import CollaborationPane from '@/components/MainHeader/CollaborationPane.vue';
+import CollaborationPane from './CollaborationPane.vue';
 import type { IUser } from '@n8n/rest-api-client/api/users';
 
 import type { RenderOptions } from '@/__tests__/render';
