+++ packages/frontend/editor-ui/src/features/collaboration/components/CollaborationPane.vue
@@ -3,7 +3,7 @@ import { onMounted, onBeforeUnmount, computed, watch } from 'vue';
 import { useDocumentVisibility } from '@vueuse/core';
 
 import { useUsersStore } from '@/stores/users.store';
-import { useCollaborationStore } from '@/stores/collaboration.store';
+import { useCollaborationStore } from '../collaboration.store';
 
 import { N8nUserStack } from '@n8n/design-system';
 const collaborationStore = useCollaborationStore();
